click on ALL CHART
... go into individual.
if you click on a value in the all chart, it will teleport you to the individual team chart for that, and on that metric.
CSS was also split off into another file.
"latestTeam" variable was removed, checking the label of the datatable is the replacer.
also changed names of the visual stuff, and made the match data below (for now) instead of on the right.